requirements.txt
google-adk>=0.1.0
google-genai>=0.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
python-multipart>=0.0.6
werkzeug>=2.3.0
gunicorn>=21.2.0
